requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
python-dateutil==2.8.2
PyPDF2==1.27.5
scikit-learn==0.24.2
plotly==5.3.1
dash==2.0.0
dash-bootstrap-components==1.0.0
pandas-datareader==0.9.0
yfinance==0.1.63
